sensors: remove airspeed confidence/data validator
The airspeed confidence could in theory detect a stuck airspeed sensor. As it is though
highly sensor specific, it is inpractical in the real world. The newer airspeed selector already
does checks for airspeed validity (checks for data dropouts, too low readings, compares to
groundspeed-windspeed) and should be the place where also additional checks reside.

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>